refactor(media): canonicalize render helpers
