+++ runtime/reply/reply_media.py
@@ -14,7 +14,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Mapping
 
-from latentsync_reply import LatentSyncReplyError, media_runtime_available, render_latentsync_video, resolve_ffmpeg_executable
+from runtime.media.latentsync_reply import LatentSyncReplyError, media_runtime_available, render_latentsync_video, resolve_ffmpeg_executable
 from runtime.media.media_paths import resolve_media_path
 from runtime.reply.reply_delivery import ReplyDeliveryPlan, plan_reply_delivery
 from voice_direction import VoicePerformancePlan
