+++ tests/media/test_minimax_music3_worker.py
@@ -14,7 +14,7 @@ from tools.minimax_profile import (
     minimax_profile_from_mapping,
 )
 from runtime.media.music_caption import render_minimax_caption
-from song_content import (
+from runtime.media.song_content import (
     PianoTexture,
     SongDynamicArc,
     SongEmotionArc,
