+++ tests/media/test_music_caption.py
@@ -9,7 +9,7 @@ from runtime.media.music_caption import (
     render_minimax_caption,
     validate_minimax_caption,
 )
-from song_content import (
+from runtime.media.song_content import (
     PianoTexture,
     SongDynamicArc,
     SongEmotionArc,
