+++ tests/media/test_portable_media_boundaries.py
@@ -12,10 +12,10 @@ import wave
 
 import pytest
 
-import latentsync_reply
+from runtime.media import latentsync_reply
 import music_reply
 from runtime.reply import reply_media
-from latentsync_reply import render_latentsync_video
+from runtime.media.latentsync_reply import render_latentsync_video
 from runtime.media.music_duration import MUSIC_DURATION_OPTIONS, normalize_music_duration
 from music_reply import MusicReplyError, render_musical_reply
 from runtime.reply.reply_delivery import (
@@ -30,7 +30,7 @@ from runtime.reply.reply_media import (
     assemble_complete_video_delivery,
     render_reply_video,
 )
-from song_content import SongContentPlan
+from runtime.media.song_content import SongContentPlan
 from tts.delivery import (
     DeliveryAudioError,
     _fit_overlong_wav,
