+++ tools/music_calibration.py
@@ -16,7 +16,7 @@ from tools.minimax_profile import (
     OFFICIAL_COMFY_MINIMAX_PROFILE,
 )
 from runtime.media.music_caption import MINIMAX_CAPTION_VERSION, render_minimax_caption
-from song_content import (
+from runtime.media.song_content import (
     PianoTexture,
     SongDynamicArc,
     SongEmotionArc,
